segfault WHOIS WHOWAS no args
SEGFAULT solved
Occurred when WHOIS, WHOWAS and NICK have no arguments.

Segfault at process_buffer solved
When the user sent a string with no positive return characters for isalnum. For example, a single empty space. Or multiple spaces and tabs.

ParseMessage improved
The following characters are removed if they are at the beginning of the input. So if the user makes “              \t           \t   \t      NICK Paul”, the Nick command is executed anyway. Special characters are not handled when they are after the command. Example:  “NICK \t     \t PAUL”, this command is not executed.
List of deleted characters: ‘ ‘ '\t' '\n' '\v' '\f' '\r'