Select error management
Change in error handling when SELECT has an error and returns -1. Previously the program would close, which would cause problems during evaluation if it happens.
After the current change, in case SELECT returns an error, we will just print the message in the STDIN and the program will continue to run.